Backend/requirements.txt
python3-ldap3
python3-flask
python3-pyotp
python3-sqlalchemy
python3-qrcode